fs-online-benchmark/locust-bechmarking/requirements.txt
locust==2.17.0
google-cloud-aiplatform==1.34.0
google-auth==2.23.2
pandas==2.1.1
boto3==1.28.53
sagemaker==2.187.0